requirements.txt
streamlit
torch
torchvision
Pillow
numpy
scikit-learn
joblib
matplotlib
seaborn